More safety checks in the super build driver script
Allow overriding the root_dir setting with an environment variable.
But provide a reasonable/existing default value if it's not
overridden.

Check for existence of key elements, and warn if they are not there.
(cmake, git, source dir...)

Add a wrapper script to make it easy to "double-click and observe"
without the window disappearing when the script is done.